auth

+++ src/Passport/RouteRegistrar.php
@@ -42,17 +42,11 @@ class RouteRegistrar {
 	 */
 	public function forAuthorization() {
 		$this->router->group(['middleware' => ['web', 'auth']], function ($router) {
-			$router->get('/authorize', [
-				'uses' => 'AuthorizationController@authorize',
-			]);
+			$router->get('/authorize', ['uses' => 'AuthorizationController@authorize']);
 
-			$router->post('/authorize', [
-				'uses' => 'ApproveAuthorizationController@approve',
-			]);
+			$router->post('/authorize', ['uses' => 'ApproveAuthorizationController@approve']);
 
-			$router->delete('/authorize', [
-				'uses' => 'DenyAuthorizationController@deny',
-			]);
+			$router->delete('/authorize', ['uses' => 'DenyAuthorizationController@deny']);
 		});
 	}
 
@@ -62,19 +56,12 @@ class RouteRegistrar {
 	 * @return void
 	 */
 	public function forAccessTokens() {
-		$this->router->post('/token', [
-			'uses' => 'AccessTokenController@issueToken',
-			'middleware' => 'throttle',
-		]);
+		$this->router->post('/token', ['uses' => 'AccessTokenController@issueToken', 'middleware' => 'throttle']);
 
 		$this->router->group(['middleware' => ['web', 'auth']], function ($router) {
-			$router->get('/tokens', [
-				'uses' => 'AuthorizedAccessTokenController@forUser',
-			]);
+			$router->get('/tokens', ['uses' => 'AuthorizedAccessTokenController@forUser']);
 
-			$router->delete('/tokens/{token_id}', [
-				'uses' => 'AuthorizedAccessTokenController@destroy',
-			]);
+			$router->delete('/tokens/{token_id}', ['uses' => 'AuthorizedAccessTokenController@destroy']);
 		});
 	}
 
@@ -84,10 +71,7 @@ class RouteRegistrar {
 	 * @return void
 	 */
 	public function forTransientTokens() {
-		$this->router->post('/token/refresh', [
-			'middleware' => ['web', 'auth'],
-			'uses' => 'TransientTokenController@refresh',
-		]);
+		$this->router->post('/token/refresh', ['middleware' => ['web', 'auth'], 'uses' => 'TransientTokenController@refresh']);
 	}
 
 	/**
@@ -97,25 +81,15 @@ class RouteRegistrar {
 	 */
 	public function forClients() {
 		$this->router->group(['middleware' => ['web', 'auth']], function ($router) {
-			$router->get('/clients', [
-				'uses' => 'ClientController@forUser',
-			]);
+			$router->get('/clients', ['uses' => 'ClientController@forUser']);
 
-			$router->post('/clients', [
-				'uses' => 'ClientController@store',
-			]);
+			$router->post('/clients', ['uses' => 'ClientController@store']);
 
-			$router->get('/clients/{client_id}', [
-				'uses' => 'ClientController@show',
-			]);
+			$router->get('/clients/{client_id}', ['uses' => 'ClientController@show']);
 
-			$router->put('/clients/{client_id}', [
-				'uses' => 'ClientController@update',
-			]);
+			$router->put('/clients/{client_id}', ['uses' => 'ClientController@update']);
 
-			$router->delete('/clients/{client_id}', [
-				'uses' => 'ClientController@destroy',
-			]);
+			$router->delete('/clients/{client_id}', ['uses' => 'ClientController@destroy']);
 		});
 	}
 
@@ -126,21 +100,13 @@ class RouteRegistrar {
 	 */
 	public function forPersonalAccessTokens() {
 		$this->router->group(['middleware' => ['web', 'auth']], function ($router) {
-			$router->get('/scopes', [
-				'uses' => 'ScopeController@all',
-			]);
+			$router->get('/scopes', ['uses' => 'ScopeController@all']);
 
-			$router->get('/personal-access-tokens', [
-				'uses' => 'PersonalAccessTokenController@forUser',
-			]);
+			$router->get('/personal-access-tokens', ['uses' => 'PersonalAccessTokenController@forUser']);
 
-			$router->post('/personal-access-tokens', [
-				'uses' => 'PersonalAccessTokenController@store',
-			]);
+			$router->post('/personal-access-tokens', ['uses' => 'PersonalAccessTokenController@store']);
 
-			$router->delete('/personal-access-tokens/{token_id}', [
-				'uses' => 'PersonalAccessTokenController@destroy',
-			]);
+			$router->delete('/personal-access-tokens/{token_id}', ['uses' => 'PersonalAccessTokenController@destroy']);
 		});
 	}
 }
